+++ test/expected/require-directory.js
@@ -1,0 +1,4 @@
+define("a/1.0.0/index", ["a/1.0.0/lib/index"], function(require, exports, module){
+require("a/1.0.0/lib/index");
+
+});